+++ src/cedalion/sigdecomp/multimodal/arc_erbm.py
@@ -1,0 +1,5 @@
+raise DeprecationWarning(
+    "cedalion.sigdecomp.multimodal.arc_erbm was renamed to "
+    "cedalion.sigdecomp.multimodal.arsc_erbm and the function arc_erbm() to "
+    "arsc_erbm()."
+)